+++ scripts/devops_tasks/test_run_samples.py
@@ -126,6 +126,8 @@ IGNORED_SAMPLES = {
         "sample_list_translations_with_filters.py",
         "sample_translation_with_custom_model.py",
         "sample_translation_with_custom_model_async.py",
+        "sample_begin_translation_with_filters.py",
+        "sample_begin_translation_with_filters_async.py"
     ],
     "azure-ai-formrecognizer": [
         "sample_manage_custom_models.py",