+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_generated/v2_1/models/_form_recognizer_client_enums.py
@@ -6,27 +6,12 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from enum import Enum, EnumMeta
+from enum import Enum
 from six import with_metaclass
-
-class _CaseInsensitiveEnumMeta(EnumMeta):
-    def __getitem__(self, name):
-        return super().__getitem__(name.upper())
-
-    def __getattr__(cls, name):
-        """Return the enum member matching `name`
-        We use __getattr__ instead of descriptors or inserting into the enum
-        class' __dict__ in order to support `name` and `value` being both
-        properties for enum members (which live in the class' __dict__) and
-        enum members themselves.
-        """
-        try:
-            return cls._member_map_[name.upper()]
-        except KeyError:
-            raise AttributeError(name)
+from azure.core import CaseInsensitiveEnumMeta
 
 
-class ContentType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class ContentType(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Content type for upload
     """
 
@@ -40,15 +25,17 @@ class ContentType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     IMAGE_PNG = "image/png"
     #: Content Type 'image/tiff'.
     IMAGE_TIFF = "image/tiff"
+    #: Content Type 'application/json'.
+    APPLICATION_JSON = "application/json"
 
-class FieldValueSelectionMark(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class FieldValueSelectionMark(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Selection mark value.
     """
 
     SELECTED = "selected"
     UNSELECTED = "unselected"
 
-class FieldValueType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class FieldValueType(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Semantic data type of the field value.
     """
 
@@ -63,14 +50,14 @@ class FieldValueType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     SELECTION_MARK = "selectionMark"
     COUNTRY_REGION = "countryRegion"
 
-class KeyValueType(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class KeyValueType(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Semantic data type of the key value element.
     """
 
     STRING = "string"
     SELECTION_MARK = "selectionMark"
 
-class Language(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class Language(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
 
     AF = "af"
     AST = "ast"
@@ -146,7 +133,7 @@ class Language(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     ZH_HANT = "zh-Hant"
     ZU = "zu"
 
-class LengthUnit(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class LengthUnit(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """The unit used by the width, height and boundingBox properties. For images, the unit is "pixel".
     For PDF, the unit is "inch".
     """
@@ -154,7 +141,7 @@ class LengthUnit(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     PIXEL = "pixel"
     INCH = "inch"
 
-class Locale(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class Locale(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
 
     EN_AU = "en-AU"
     EN_CA = "en-CA"
@@ -162,7 +149,7 @@ class Locale(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     EN_IN = "en-IN"
     EN_US = "en-US"
 
-class ModelStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class ModelStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Status of the model.
     """
 
@@ -170,7 +157,7 @@ class ModelStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     READY = "ready"
     INVALID = "invalid"
 
-class OperationStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class OperationStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Status of the queued operation.
     """
 
@@ -179,26 +166,26 @@ class OperationStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
     SUCCEEDED = "succeeded"
     FAILED = "failed"
 
-class ReadingOrder(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class ReadingOrder(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
 
     BASIC = "basic"
     NATURAL = "natural"
 
-class SelectionMarkState(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class SelectionMarkState(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """State of the selection mark.
     """
 
     SELECTED = "selected"
     UNSELECTED = "unselected"
 
-class TextStyle(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class TextStyle(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """The text line style name, including handwriting and other.
     """
 
     OTHER = "other"
     HANDWRITING = "handwriting"
 
-class TrainStatus(with_metaclass(_CaseInsensitiveEnumMeta, str, Enum)):
+class TrainStatus(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
     """Status of the training operation.
     """