+++ sdk/search/azure-search-documents/tests/consts.py
@@ -1,9 +1,0 @@
-# ------------------------------------
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-# -----------------------------------
-
-TEST_SERVICE_NAME = "test-service-name"
-SERVICE_URL = "https://{}.search.windows.net/indexes?api-version=2021-04-30-Preview".format(
-    TEST_SERVICE_NAME
-)